pipico/graph_multi_data_points/plot/requirements.txt
pyserial>=3.5
pyqtgraph>=0.13.3
PyQt5>=5.15